move config resolver
