+++ dfine_seg/config/__init__.py
@@ -1,0 +1,1 @@
+"""Packaged config: the `dfine init` template and the code that locates a live one."""
