+++ dfine_seg/dl/infer.py
@@ -8,7 +8,7 @@ from loguru import logger
 from omegaconf import DictConfig
 from tqdm import tqdm
 
-from dfine_seg._config import CONFIG_NAME, config_dir
+from dfine_seg.config.resolve import CONFIG_NAME, config_dir
 from dfine_seg.dl.dataset import read_image_hwc
 from dfine_seg.dl.utils import (
     abs_xyxy_to_norm_xywh,
