+++ dfine_seg/dl/trt_int8.py
@@ -23,7 +23,7 @@ from loguru import logger
 from omegaconf import DictConfig
 from tqdm import tqdm
 
-from dfine_seg._config import CONFIG_NAME, config_dir
+from dfine_seg.config.resolve import CONFIG_NAME, config_dir
 from dfine_seg.dl.dataset import Loader
 from dfine_seg.dl.train import Trainer
 from dfine_seg.dl.utils import get_latest_experiment_name
