+++ dfine_seg/etl/clean_csv.py
@@ -5,7 +5,7 @@ import hydra
 import pandas as pd
 from omegaconf import DictConfig
 
-from dfine_seg._config import CONFIG_NAME, config_dir
+from dfine_seg.config.resolve import CONFIG_NAME, config_dir
 
 
 @hydra.main(version_base=None, config_path=config_dir(), config_name=CONFIG_NAME)
