+++ dfine_seg/etl/split.py
@@ -8,7 +8,7 @@ from loguru import logger
 from omegaconf import DictConfig
 from sklearn.model_selection import train_test_split
 
-from dfine_seg._config import CONFIG_NAME, config_dir
+from dfine_seg.config.resolve import CONFIG_NAME, config_dir
 
 COCO_SRC = "coco.json"  # single-file COCO annotations, split into train/val(/test).json
